+++ app/components/UI/AccountSelectorList/AccountSelectorList.tsx
@@ -39,6 +39,7 @@ import { AccountListBottomSheetSelectorsIDs } from '../../../../e2e/selectors/wa
 import { WalletViewSelectorsIDs } from '../../../../e2e/selectors/wallet/WalletView.selectors';
 import { RootState } from '../../../reducers';
 import { ACCOUNT_SELECTOR_LIST_TESTID } from './AccountSelectorList.constants';
+import { toHex } from '@metamask/controller-utils';
 
 const AccountSelectorList = ({
   onSelectAccount,
@@ -167,10 +168,10 @@ const AccountSelectorList = ({
                   removedAddress: address,
                   nextActiveAddress,
                 });
-                await Engine.context.KeyringController.removeAccount(address);
                 // Revocation of accounts from PermissionController is needed whenever accounts are removed.
                 // If there is an instance where this is not the case, this logic will need to be updated.
-                removeAccountsFromPermissions([address]);
+                removeAccountsFromPermissions([toHex(address)]);
+                await Engine.context.KeyringController.removeAccount(address);
               });
             },
           },