+++ app/components/Views/Settings/PermissionsSettings/PermissionsManager.tsx
@@ -27,7 +27,15 @@ import {
 import {
   PermissionControllerState,
   PermissionConstraint,
+  Caveat,
 } from '@metamask/permission-controller';
+import {
+  Caip25CaveatType,
+  Caip25EndowmentPermissionName,
+  getEthAccounts,
+  getPermittedEthChainIds,
+} from '@metamask/chain-agnostic-permission';
+import { Json } from '@metamask/utils';
 
 interface SDKSessionsManagerProps {
   navigation: NavigationProp<ParamListBase>;
@@ -91,16 +99,33 @@ const PermissionsManager = (props: SDKSessionsManagerProps) => {
     });
 
     const mappedInAppBrowserPermissions: PermissionListItemViewModel[] =
-      inAppBrowserSubjects.map((subject) => ({
-        dappLogoUrl: '',
-        dappHostName: subject.origin,
-        numberOfAccountPermissions:
-          subject.permissions?.eth_accounts?.caveats?.[0]?.value?.length ?? 0,
-        numberOfNetworkPermissions:
-          subject.permissions?.['endowment:permitted-chains']?.caveats?.[0]
-            ?.value?.length ?? 0,
-        permissionSource: PermissionSource.MetaMaskBrowser,
-      }));
+      inAppBrowserSubjects.map((subject) => {
+        const caip25CaveatValue = subject.permissions?.[
+          Caip25EndowmentPermissionName
+        ]?.caveats?.find(
+          (caveat: { type: string; value: Caveat<string, Json> }) =>
+            caveat.type === Caip25CaveatType,
+        )?.value ?? {
+          optionalScopes: {
+            'wallet:eip155': { accounts: [] },
+          },
+          requiredScopes: {},
+          sessionProperties: {},
+          isMultichainOrigin: false,
+        };
+
+        return {
+          dappLogoUrl: '',
+          dappHostName: subject.origin,
+          numberOfAccountPermissions: caip25CaveatValue
+            ? getEthAccounts(caip25CaveatValue).length
+            : 0,
+          numberOfNetworkPermissions: caip25CaveatValue
+            ? getPermittedEthChainIds(caip25CaveatValue).length
+            : 0,
+          permissionSource: PermissionSource.MetaMaskBrowser,
+        };
+      });
 
     const mappedPermissions: PermissionListItemViewModel[] = [
       ...mappedInAppBrowserPermissions,