+++ app/core/RPCMethods/createEip1193MethodMiddleware/index.test.js
@@ -1,60 +1,53 @@
+import { JsonRpcEngine } from '@metamask/json-rpc-engine';
 import {
-  JsonRpcEngine,
-  JsonRpcEngineEndCallback,
-  JsonRpcEngineNextCallback,
-} from '@metamask/json-rpc-engine';
-import {
-  Json,
-  JsonRpcParams,
-  JsonRpcRequest,
-  PendingJsonRpcResponse,
   assertIsJsonRpcFailure,
   assertIsJsonRpcSuccess,
 } from '@metamask/utils';
-import createLegacyMethodMiddleware from '.';
+import { createEip1193MethodMiddleware } from '.';
 
-jest.mock('./util', () => {
-  const getHandler = () => ({
-    implementation: (
-      req: JsonRpcRequest<JsonRpcParams>,
-      res: PendingJsonRpcResponse<Json>,
-      _next: JsonRpcEngineNextCallback,
-      end: JsonRpcEngineEndCallback,
-      // TODO: Replace "any" with type
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      hooks: Record<string, any>,
-    ) => {
-      if (Array.isArray(req.params)) {
-        switch (req.params[0]) {
-          case 1:
-            res.result = hooks.hook1();
-
-            break;
-          case 2:
-            res.result = hooks.hook2();
-
-            break;
-          case 3:
-            return end(new Error('test error'));
-          case 4:
-            throw new Error('test error');
-          case 5:
-            // eslint-disable-next-line no-throw-literal
-            throw 'foo';
-          default:
-            throw new Error(`unexpected param "${req.params[0]}"`);
-        }
+const getHandler = () => ({
+  implementation: (req, res, _next, end, hooks) => {
+    if (Array.isArray(req.params)) {
+      switch (req.params[0]) {
+        case 1:
+          res.result = hooks.hook1();
+          break;
+        case 2:
+          res.result = hooks.hook2();
+          break;
+        case 3:
+          return end(new Error('test error'));
+        case 4:
+          throw new Error('test error');
+        case 5:
+          // eslint-disable-next-line no-throw-literal
+          throw 'foo';
+        default:
+          throw new Error(`unexpected param "${req.params[0]}"`);
       }
-      return end();
-    },
-    hookNames: { hook1: true, hook2: true },
-    methodNames: ['method1', 'method2'],
-  });
-
-  return [getHandler()];
+    }
+    return end();
+  },
+  hookNames: { hook1: true, hook2: true },
+  methodNames: ['method1', 'method2'],
 });
 
-describe('createLegacyMethodMiddleware', () => {
+jest.mock('@metamask/permission-controller', () => ({
+  ...jest.requireActual('@metamask/permission-controller'),
+}));
+
+jest.mock('@metamask/eip1193-permission-middleware', () => ({
+  getPermissionsHandler: getHandler(),
+  requestPermissionsHandler: getHandler(),
+  revokePermissionsHandler: getHandler(),
+}));
+
+jest.mock('../handlers', () => ({
+  localHandlers: [getHandler()],
+  eip1193OnlyHandlers: [getHandler()],
+}));
+
+describe('createEip1193MethodMiddleware', () => {
   const method1 = 'method1';
 
   const getDefaultHooks = () => ({
@@ -63,14 +56,14 @@ describe('createLegacyMethodMiddleware', () => {
   });
 
   it('should return a function', () => {
-    const middleware = createLegacyMethodMiddleware(getDefaultHooks());
+    const middleware = createEip1193MethodMiddleware(getDefaultHooks());
     expect(typeof middleware).toBe('function');
   });
 
   it('should throw an error if a required hook is missing', () => {
     const hooks = { hook1: () => 42 };
 
-    expect(() => createLegacyMethodMiddleware(hooks)).toThrow(
+    expect(() => createEip1193MethodMiddleware(hooks)).toThrow(
       'Missing expected hooks',
     );
   });
@@ -81,13 +74,13 @@ describe('createLegacyMethodMiddleware', () => {
       extraneousHook: () => 100,
     };
 
-    expect(() => createLegacyMethodMiddleware(hooks)).toThrow(
+    expect(() => createEip1193MethodMiddleware(hooks)).toThrow(
       'Received unexpected hooks',
     );
   });
 
   it('should call the handler for the matching method (uses hook1)', async () => {
-    const middleware = createLegacyMethodMiddleware(getDefaultHooks());
+    const middleware = createEip1193MethodMiddleware(getDefaultHooks());
     const engine = new JsonRpcEngine();
     engine.push(middleware);
 
@@ -103,7 +96,7 @@ describe('createLegacyMethodMiddleware', () => {
   });
 
   it('should call the handler for the matching method (uses hook2)', async () => {
-    const middleware = createLegacyMethodMiddleware(getDefaultHooks());
+    const middleware = createEip1193MethodMiddleware(getDefaultHooks());
     const engine = new JsonRpcEngine();
     engine.push(middleware);
 
@@ -119,7 +112,7 @@ describe('createLegacyMethodMiddleware', () => {
   });
 
   it('should not call the handler for a non-matching method', async () => {
-    const middleware = createLegacyMethodMiddleware(getDefaultHooks());
+    const middleware = createEip1193MethodMiddleware(getDefaultHooks());
     const engine = new JsonRpcEngine();
     engine.push(middleware);
 
@@ -138,7 +131,7 @@ describe('createLegacyMethodMiddleware', () => {
   });
 
   it('should handle errors returned by the implementation', async () => {
-    const middleware = createLegacyMethodMiddleware(getDefaultHooks());
+    const middleware = createEip1193MethodMiddleware(getDefaultHooks());
     const engine = new JsonRpcEngine();
     engine.push(middleware);
 
@@ -150,15 +143,12 @@ describe('createLegacyMethodMiddleware', () => {
     });
     assertIsJsonRpcFailure(response);
 
-    // Type assertion for the error not having cause object
-    const errorData = response.error.data as { cause?: Error };
-
-    expect(errorData.cause?.message).toBe('test error');
     expect(response.error.message).toBe('test error');
+    expect(response.error.data.cause.message).toBe('test error');
   });
 
   it('should handle errors thrown by the implementation', async () => {
-    const middleware = createLegacyMethodMiddleware(getDefaultHooks());
+    const middleware = createEip1193MethodMiddleware(getDefaultHooks());
     const engine = new JsonRpcEngine();
     engine.push(middleware);
 
@@ -170,15 +160,12 @@ describe('createLegacyMethodMiddleware', () => {
     });
     assertIsJsonRpcFailure(response);
 
-    // Type assertion for the error not having cause object
-    const errorData = response.error.data as { cause?: Error };
-
-    expect(errorData.cause?.message).toBe('test error');
     expect(response.error.message).toBe('test error');
+    expect(response.error.data.cause.message).toBe('test error');
   });
 
   it('should handle non-errors thrown by the implementation', async () => {
-    const middleware = createLegacyMethodMiddleware(getDefaultHooks());
+    const middleware = createEip1193MethodMiddleware(getDefaultHooks());
     const engine = new JsonRpcEngine();
     engine.push(middleware);