+++ app/core/RPCMethods/eth_accounts/types.ts
@@ -2,5 +2,5 @@
  * Implementations used by eth_accounts handler
  */
 export interface EthAccountHooks {
-  getAccounts: () => Promise<string[]>;
+  getAccounts: () => string[];
 }