+++ app/core/RPCMethods/lib/ethereum-chain-utils.test.ts
@@ -3,9 +3,27 @@ import MetaMetrics from '../../Analytics/MetaMetrics';
 import { MetricsEventBuilder } from '../../Analytics/MetricsEventBuilder';
 import { MetaMetricsEvents } from '../../Analytics';
 import { switchToNetwork } from './ethereum-chain-utils';
+import { getDefaultCaip25CaveatValue } from '../../Permissions';
 
 jest.mock('../../Analytics/MetaMetrics');
 jest.mock('../../Analytics/MetricsEventBuilder');
+jest.mock('../../../core/Permissions', () => ({
+  ...jest.requireActual('../../../core/Permissions'),
+  getPermittedAccounts: jest.fn().mockReturnValue([]),
+}));
+jest.mock('../../Engine', () => ({
+  context: {
+    MultichainNetworkController: {
+      setActiveNetwork: jest.fn(),
+    },
+    PermissionController: {
+      grantPermissionsIncremental: jest.fn(),
+    },
+    SelectedNetworkController: {
+      setNetworkClientIdForDomain: jest.fn(),
+    },
+  },
+}));
 
 describe('switchToNetwork', () => {
   it('tracks the network switch event', async () => {
@@ -21,6 +39,16 @@ describe('switchToNetwork', () => {
       build: jest.fn().mockReturnValue(mockMetricsBuilderBuild),
     });
 
+    const mockHooks = {
+      getCaveat: jest
+        .fn()
+        .mockReturnValue({ value: getDefaultCaip25CaveatValue() }),
+      requestPermittedChainsPermissionIncrementalForOrigin: jest.fn(),
+      hasApprovalRequestsForOrigin: jest.fn(),
+      toNetworkConfiguration: jest.fn(),
+      fromNetworkConfiguration: jest.fn(),
+    };
+
     const chainId = '0x1';
     const {
       selectedNetworkClientId: networkClientId,
@@ -32,18 +60,6 @@ describe('switchToNetwork', () => {
       ticker: 'ETH',
     });
 
-    const mockMultichainNetworkController = {
-      setActiveNetwork: jest.fn(),
-    };
-
-    const mockPermissionController = {
-      getCaveat: jest.fn(),
-      hasPermission: jest.fn().mockReturnValue(true),
-      grantPermissionsIncremental: jest.fn(),
-    };
-
-    const mockSelectedNetworkController = {};
-
     const requestUserApproval = jest.fn();
     const analytics = {
       test: 'test',
@@ -54,15 +70,11 @@ describe('switchToNetwork', () => {
     await switchToNetwork({
       network: [networkClientId, network],
       chainId,
-      controllers: {
-        MultichainNetworkController: mockMultichainNetworkController,
-        PermissionController: mockPermissionController,
-        SelectedNetworkController: mockSelectedNetworkController,
-      },
       requestUserApproval,
       analytics,
       origin,
       isAddNetworkFlow,
+      hooks: mockHooks,
     });
 
     expect(MetricsEventBuilder.createEventBuilder).toHaveBeenCalledWith(