+++ app/core/__mocks__/MockedEngine.ts
@@ -6,11 +6,61 @@ import { MOCK_KEYRING_CONTROLLER_STATE } from '../../util/test/keyringController
 
 export const mockedEngine = {
   init: () => Engine.init({}),
+  controllerMessenger: {
+    subscribeOnceIf: jest.fn(),
+    subscribe: jest.fn(),
+    call: jest.fn().mockImplementation((method) => {
+      if (method === 'SelectedNetworkController:getNetworkClientIdForDomain') {
+        return 'mainnet';
+      }
+
+      if (method === 'NetworkController:getNetworkClientById') {
+        return {
+          configuration: {
+            chainId: '0x1',
+            ticker: 'ETH',
+          },
+        };
+      }
+    }),
+  },
+  datamodel: {
+    state: { PreferencesController: { selectedAddress: '' } },
+  },
   context: {
-    KeyringController: {
-      state: MOCK_KEYRING_CONTROLLER_STATE,
+    AccountsController: {
+      listAccounts: jest.fn(),
+      getSelectedAccount: jest.fn(),
     },
+    AccountTrackerController: {
+      state: {
+        accounts: {},
+      },
+    },
+    ApprovalController: {
+      addAndShowApprovalRequest: jest.fn(),
+    },
+    PermissionController: {
+      createPermissionMiddleware: jest.fn(),
+      requestPermissions: jest.fn(),
+      getCaveat: jest.fn(),
+      updateCaveat: jest.fn(),
+      revokePermission: jest.fn(),
+      revokePermissions: jest.fn(),
+      getPermissions: jest.fn(),
+      hasPermissions: jest.fn(),
+      hasPermission: jest.fn(),
+      executeRestrictedMethod: jest.fn(),
+      state: {
+        subjects: {},
+      },
+    },
+    SelectedNetworkController: {
+      getProviderAndBlockTracker: jest.fn(),
+    },
+    KeyringController: MOCK_KEYRING_CONTROLLER_STATE,
     NetworkController: {
+      getNetworkConfigurationByChainId: jest.fn(),
       getNetworkClientById: (networkClientId: NetworkClientId) => {
         if (networkClientId === 'linea_goerli') {
           return {
@@ -44,6 +94,7 @@ export const mockedEngine = {
     },
   },
   hasFunds: jest.fn(),
+  getCaip25PermissionFromLegacyPermissions: jest.fn(),
 };
 
 export default mockedEngine;